Make use of DiscoveredServersHandler to avoid repeated polling
Rebased master to make use of the updated vendor directory. The
Go client has now a new handler that is notified when a new server
is discovered. Make use of that and replace repeat polling and
array comparison and simply send the list of channels when this
handler is invoked.